+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 * [#219](https://github.com/TOMO-CAT/xmake/issues/219): 支持配置自动删除 cache 和 packages 的轮转时间
   * filelock 里增加时间戳和格式化时间戳字段, 用于判断是否长时间没人用
   * package 文件锁迁移到 `~/.xmake/filelock` 目录下
+  * 每天清理一次, 日志会写到 xmake tmp 目录下的 cleaner.log 里
 
 ### Bugs 修复
 
 * [#241](https://github.com/TOMO-CAT/xmake/issues/241): 修复 package fullpath 导致的 linkgroups 等配置的 bug
+* [#244](https://github.com/TOMO-CAT/xmake/issues/244): 修复并发安装同一个 package 会 panic 的问题
 
 ## v3.0.7